requirements.txt
altgraph==0.17.4
ghostscript==0.7
lxml==5.3.0
numpy==2.1.1
opencv-python==4.10.0.84
packaging==24.1
pefile==2024.8.26
pillow==10.4.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyPDF2==3.0.1
PyQt6==6.7.1
PyQt6-Qt6==6.7.2
PyQt6_sip==13.8.0
pytesseract==0.3.13
python-docx==1.1.2
pywin32-ctypes==0.2.3
pyzbar==0.1.9
setuptools==75.1.0
typing_extensions==4.12.2
